+++ Lorr/Engine/ECS/Components/PhysicsComponent.hh
@@ -1,0 +1,17 @@
+//
+// Created on Thursday 7th October 2021 by e-erdal
+//
+
+#pragma once
+
+#include "Engine/App/Engine.hh"
+#include "Engine/ECS/Entity.hh"
+
+namespace Lorr::Component
+{
+    struct Physics
+    {
+        b2Body *pBody = 0;
+    };
+
+}  // namespace Lorr::Component