+++ Lorr/Engine/ECS/Components/TransformComponent.hh
@@ -10,11 +10,29 @@ namespace Lorr::Component
 {
     struct Transform
     {
+        Transform(glm::vec3 pos, glm::vec3 size = {}, float rotation = 0.f) : Position(pos), Size(size)
+        {
+            Rotation = glm::radians(rotation);
+            SetOriginTopLeft();
+        }
+
         glm::vec3 Position;
         glm::vec3 Size;
         float Rotation = 0.f;
+        glm::vec3 RotationOrigin;
 
         glm::mat4 Matrix;
+
+        void SetOriginCenter()
+        {
+            RotationOrigin = Size / 2.f;
+        }
+
+        void SetOriginTopLeft()
+        {
+            RotationOrigin.x = 0;
+            RotationOrigin.y = 0;
+        }
     };
 
 }  // namespace Lorr::Component