+++ Lorr/Engine/ECS/Systems/TransformSystem.cc
@@ -6,17 +6,20 @@ namespace Lorr::System
 {
     void TransformSystem::Tick(float deltaTime)
     {
-        glm::mat4 matrix(1.0f);
         auto view = m_pRegistry->view<Component::Transform>();
 
         for (auto entity : view)
         {
             auto &transform = view.get<Component::Transform>(entity);
 
-            Math::SetPos(matrix, transform.Position);
-            matrix = glm::rotate(matrix, glm::radians(transform.Rotation), glm::vec3(0.f, 0.f, 1.f));
-            Math::SetSize(matrix, transform.Size);
-            transform.Matrix = matrix;
+            transform.Matrix = glm::mat4(1.0f);
+            transform.Matrix = glm::translate(transform.Matrix, transform.Position);
+
+            transform.Matrix = glm::translate(transform.Matrix, transform.RotationOrigin);
+            transform.Matrix = glm::rotate(transform.Matrix, transform.Rotation, glm::vec3(0.0f, 0.0f, 1.f));
+            transform.Matrix = glm::translate(transform.Matrix, -transform.RotationOrigin);
+            
+            transform.Matrix = glm::scale(transform.Matrix, transform.Size);
         }
     }